Reorganized common trade token modal components
