+++ packages/commonwealth/client/scripts/views/components/sidebar/CommunitySection/TokenTradeWidget/TokenTradeWidget.tsx
@@ -5,11 +5,10 @@ import { calculateTokenPricing } from 'helpers/launchpad';
 import React, { useState } from 'react';
 import app from 'state';
 import { useFetchTokenUsdRateQuery } from 'state/api/communityStake';
-import TradeTokenModal from 'views/modals/TradeTokenModel';
-import {
+import TradeTokenModal, {
   TokenWithCommunity,
   TradingMode,
-} from 'views/modals/TradeTokenModel/CommonTradeModal/TradeTokenForm';
+} from 'views/modals/TradeTokenModel';
 import { z } from 'zod';
 import { CWDivider } from '../../../component_kit/cw_divider';
 import { CWIconButton } from '../../../component_kit/cw_icon_button';
