+++ packages/commonwealth/client/scripts/views/modals/TradeTokenModel/CommonTradeModal/CommonTradeTokenForm/CommonTradeTokenForm.scss
@@ -1,6 +1,6 @@
 @import '../../../../../styles/shared.scss';
 
-.TradeTokenForm {
+.CommonTradeTokenForm {
   display: flex;
   flex-direction: column;
   gap: 12px;
