+++ packages/commonwealth/client/scripts/views/modals/TradeTokenModel/CommonTradeModal/CommonTradeTokenForm/ReceiptDetails/ReceiptDetails.scss
