+++ packages/commonwealth/client/scripts/views/modals/TradeTokenModel/CommonTradeModal/CommonTradeTokenForm/index.ts
@@ -1,0 +1,6 @@
+import CommonTradeTokenForm from './CommonTradeTokenForm';
+import useCommonTradeTokenForm from './useCommonTradeTokenForm';
+export * from './types';
+export * from './useCommonTradeTokenForm';
+export { useCommonTradeTokenForm };
+export default CommonTradeTokenForm;
