--- packages/commonwealth/client/scripts/views/modals/TradeTokenModel/CommonTradeModal/TradeTokenForm/index.ts
@@ -1,6 +1,0 @@
-import TradeTokenForm from './TradeTokenForm';
-import useTradeTokenForm from './useTradeTokenForm';
-export * from './types';
-export * from './useTradeTokenForm';
-export { useTradeTokenForm };
-export default TradeTokenForm;
