+++ packages/commonwealth/client/scripts/views/modals/TradeTokenModel/index.ts
@@ -1,3 +1,4 @@
 import TradeTokenModal from './TradeTokenModal';
+export * from './types';
 
 export default TradeTokenModal;
