+++ packages/commonwealth/client/scripts/views/pages/Communities/TokensList/TokensList.tsx
@@ -11,8 +11,7 @@ import { useFetchTokensQuery } from 'state/api/tokens';
 import { CWText } from 'views/components/component_kit/cw_text';
 import { CWButton } from 'views/components/component_kit/new_designs/CWButton';
 import CWCircleMultiplySpinner from 'views/components/component_kit/new_designs/CWCircleMultiplySpinner';
-import TradeTokenModal from 'views/modals/TradeTokenModel';
-import { TradingMode } from 'views/modals/TradeTokenModel/CommonTradeModal/TradeTokenForm';
+import TradeTokenModal, { TradingMode } from 'views/modals/TradeTokenModel';
 import { z } from 'zod';
 import TokenCard from '../../../components/TokenCard';
 import {
